Removed maxDistance bound on my bond search alogrithm.  Once I figure out why that bound wasn't correct, I'll readd it.  I aslo made changes to the atom class.  If the element symbol isn't in the pdb file, it uses the atom name to determine the element symbol.

+++ trunk/osgPlugins/pdb/Cylinder.h
@@ -21,7 +21,7 @@ class Cylinder
 {
 public:
   Cylinder(const osg::Vec3&, const osg::Vec3&);
-  osg::Geometry * getGeometry(osg::Material *, float) const;
+  osg::Geometry * getGeometry(osg::Material *, unsigned int) const;
 private:
   Cylinder();
   float _getDist(const osg::Vec3&, const osg::Vec3&) const;
